libm : new ldexp frexp and log2 implementation

+++ libm/generic/log2f.c
@@ -2,7 +2,7 @@
 #include <errno.h>
 #include "internal.h"
 
-const double c[9] = {
+const float c[9] = {
 	1.4426950408889634, // 1/ln(2)
 	-0.7213475204444817,
 	 0.4808983469629871,
@@ -14,7 +14,7 @@ const double c[9] = {
 	 0.1603052267555543,
 };
 
-double log2(double x) {
+float log2f(float x) {
 	if (isinf(x) || isnan(x)) {
 		return x;
 	}
@@ -28,16 +28,16 @@ double log2(double x) {
 		}
 	}
 	int exp;
-	double m = frexp(x, &exp);
+	float m = frexpf(x, &exp);
 
 	// get closer to 0
-	double z = m - 1.0;
+	float z = m - 1.0;
 
 	// 9 degree minimax polyminal
-	double log2_m = c[8];
+	float log2_m = c[8];
 	for (int i=7; i>=0; i--) {
 		log2_m = log2_m * z + c[i];
 	}
 	log2_m *= z;
-	return (double)exp + log2_m;
+	return (float)exp + log2_m;
 }
